@@ -1,47 +1,47 @@
 {
-  "_args": [
-    [
-      "dotenv@6.2.0",
-      "D:\\Sample Projects\\Node\\InventorySystem"
-    ]
-  ],
-  "_from": "dotenv@6.2.0",
-  "_id": "dotenv@6.2.0",
+  "_from": "dotenv",
+  "_id": "dotenv@8.2.0",
   "_inBundle": false,
-  "_integrity": "sha512-HygQCKUBSFl8wKQZBSemMywRWcEDNidvNbjGVyZu3nbZ8qq9ubiPoGLMdRDpfSrpkkm9BXYFkpKxxFX38o/76w==",
+  "_integrity": "sha512-8sJ78ElpbDJBHNeBzUbUVLsqKdccaa/BXF1uPTw3GrvQTBgrQrtObr2mUrE38vzYd8cEv+m/JBfDLioYcfXoaw==",
   "_location": "/dotenv",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "tag",
     "registry": true,
-    "raw": "dotenv@6.2.0",
+    "raw": "dotenv",
     "name": "dotenv",
     "escapedName": "dotenv",
-    "rawSpec": "6.2.0",
+    "rawSpec": "",
     "saveSpec": null,
-    "fetchSpec": "6.2.0"
+    "fetchSpec": "latest"
   },
   "_requiredBy": [
-    "/typeorm"
+    "#USER",
+    "/"
   ],
-  "_resolved": "https://registry.npmjs.org/dotenv/-/dotenv-6.2.0.tgz",
-  "_spec": "6.2.0",
+  "_resolved": "https://registry.npmjs.org/dotenv/-/dotenv-8.2.0.tgz",
+  "_shasum": "97e619259ada750eea3e4ea3e26bceea5424b16a",
+  "_spec": "dotenv",
   "_where": "D:\\Sample Projects\\Node\\InventorySystem",
   "bugs": {
     "url": "https://github.com/motdotla/dotenv/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {},
+  "deprecated": false,
   "description": "Loads environment variables from .env file",
   "devDependencies": {
-    "decache": "^4.5.0",
-    "flow-bin": "^0.84.0",
-    "sinon": "^6.3.5",
-    "standard": "^12.0.1",
-    "standard-markdown": "^5.0.1",
-    "tap": "^12.0.1"
+    "decache": "^4.5.1",
+    "dtslint": "^0.9.8",
+    "flow-bin": "^0.109.0",
+    "sinon": "^7.5.0",
+    "standard": "^13.1.0",
+    "standard-markdown": "^5.1.0",
+    "standard-version": "^7.0.0",
+    "tap": "^14.7.0"
   },
   "engines": {
-    "node": ">=6"
+    "node": ">=8"
   },
   "homepage": "https://github.com/motdotla/dotenv#readme",
   "keywords": [
@@ -61,10 +61,13 @@
     "url": "git://github.com/motdotla/dotenv.git"
   },
   "scripts": {
+    "dtslint": "dtslint types",
     "flow": "flow",
     "lint": "standard",
     "postlint": "standard-markdown",
-    "pretest": "npm run lint",
+    "prerelease": "npm test",
+    "pretest": "npm run lint && npm run dtslint",
+    "release": "standard-version",
     "test": "tap tests/*.js --100"
   },
   "standard": {
@@ -72,5 +75,6 @@
       "flow-typed/"
     ]
   },
-  "version": "6.2.0"
+  "types": "types",
+  "version": "8.2.0"
 }
