@@ -1,12 +1,13 @@
+import { ILoggerService } from './../interfaces/ilogger.service';
 import * as express from 'express';
 import { interfaces, controller, httpGet, request, response, httpPost, httpPut } from 'inversify-express-utils';
 import { inject } from 'inversify';
 import TYPES from '../ioc/types';
 import { Product } from '../entities/product.entity';
 import { Cart } from '../entities/cart.entity';
-import { IProductRepository } from './../repositories/iproduct.repository';
+import { IProductRepository } from '../interfaces/iproduct.repository';
 import { User } from './../entities/user.entity';
-import { ICartRepository } from './../repositories/icart.repository';
+import { ICartRepository } from '../interfaces/icart.repository';
 
 /**
  * Cart Controller
@@ -16,16 +17,20 @@ export class CartController implements interfaces.Controller {
 
     private cartRepository : ICartRepository;
     private productRepository : IProductRepository;
+    private loggerService : ILoggerService;
 
     /**
      * Creates an instance of cart controller.
      * @param cartRepository 
      * @param productRepository 
+     * @param loggerService
      */
     constructor(@inject(TYPES.ICartRepository) cartRepository : ICartRepository,
-                @inject(TYPES.IProductRepository) productRepository : IProductRepository) {
+                @inject(TYPES.IProductRepository) productRepository : IProductRepository,
+                @inject(TYPES.ILoggerService) loggerService : ILoggerService) {
         this.cartRepository = cartRepository;
         this.productRepository = productRepository;
+        this.loggerService = loggerService;
     }
 
     /**
@@ -43,6 +48,7 @@ export class CartController implements interfaces.Controller {
                 res.status(200).json(cart);
             }
         } catch(error) {
+            this.loggerService.logError(JSON.stringify(error));
             res.status(400).json(error);
         }
     }
@@ -71,10 +77,15 @@ export class CartController implements interfaces.Controller {
             {
                 const result = await this.cartRepository.create(cart);
                 res.status(200).json(result);
-            } else {
-                res.status(400).json('Product stock is very less');
+            }
+            else {
+                let errorMessage = 'Product stock is very less';
+                this.loggerService.logError(errorMessage);
+                res.status(400).json(errorMessage);
             }
         } catch(error) {
+            console.log(error);
+            this.loggerService.logError(JSON.stringify(error));
             res.status(400).json(error);
         }
     }
@@ -99,6 +110,8 @@ export class CartController implements interfaces.Controller {
                 res.status(400).json('Something went wrong');
             }
         } catch(error) {
+            console.log(error);
+            this.loggerService.logError(JSON.stringify(error));
             res.status(400).json(error);
         }
     }
