@@ -1,11 +1,13 @@
+import { ILoggerService } from './../interfaces/ilogger.service';
+import { IEventPublishService } from '../interfaces/ievent-publish.service';
 import * as express from 'express';
 import { interfaces, controller, httpGet, request, response, httpPost } from 'inversify-express-utils';
 import { inject } from 'inversify';
 import TYPES from '../ioc/types';
 import { User } from './../entities/user.entity';
 import { Order } from './../entities/order.entity';
-import { IOrderRepository } from './../repositories/iorder.repository';
-import { ICartRepository } from './../repositories/icart.repository';
+import { IOrderRepository } from '../interfaces/iorder.repository';
+import { ICartRepository } from '../interfaces/icart.repository';
 
 /**
  * Order Controller
@@ -15,6 +17,8 @@ export class OrderController implements interfaces.Controller {
 
     private cartRepository : ICartRepository;
     private orderRepository : IOrderRepository;
+    private eventPublishService : IEventPublishService;
+    private loggerService : ILoggerService;
 
     /**
      * Creates an instance of order controller.
@@ -22,9 +26,13 @@ export class OrderController implements interfaces.Controller {
      * @param orderRepository 
      */
     constructor(@inject(TYPES.ICartRepository) cartRepository : ICartRepository,
-                @inject(TYPES.IOrderRepository) orderRepository : IOrderRepository) {
+                @inject(TYPES.IOrderRepository) orderRepository : IOrderRepository,
+                @inject(TYPES.IEventPublishService) eventPublishService : IEventPublishService,
+                @inject(TYPES.ILoggerService) loggerService : ILoggerService) {
         this.cartRepository = cartRepository;
         this.orderRepository = orderRepository;
+        this.eventPublishService = eventPublishService;
+        this.loggerService = loggerService;
     }
 
     /**
@@ -42,6 +50,7 @@ export class OrderController implements interfaces.Controller {
                 res.status(200).json(order);
             }
         } catch(error) {
+            this.loggerService.logError(JSON.stringify(error));
             res.status(400).json(error);
         }
     }
@@ -61,6 +70,7 @@ export class OrderController implements interfaces.Controller {
                 res.status(200).json(orders);
             }
         } catch(error) {
+            this.loggerService.logError(JSON.stringify(error));
             res.status(400).json(error);
         }
     }
@@ -77,9 +87,16 @@ export class OrderController implements interfaces.Controller {
             let user = new User();
             user.UserId = req.body.userId;
             order.User = user;
-            const result = await this.orderRepository.create(order);
+            const result : {
+                OrderId:number,
+                Total:number,
+                OrderDate:Date
+            } = await this.orderRepository.create(order);
+            const newOrder = await this.orderRepository.findById(result[0].OrderId);
+            this.eventPublishService.publish(newOrder);
             res.status(200).json(result);
         } catch(error) {
+            this.loggerService.logError(JSON.stringify(error));
             res.status(400).json(error);
         }
     }
